tests/requirements.txt
django<2.1
django-rest-framework=3.7.7
django-environ==0.4.4
pytest==3.4.0
pytest-django==3.1.2
pytest-env==0.6.2
pytest-pythonpath==0.7.2
pytest-cov==2.5.1
celery==4.1.0
redis==2.10.6
requests==2.18.4
requests-toolbelt==0.8.0
responses==0.8.1
#raven==6.5.0
model_mommy==1.5.1
ipdb==0.10.3
git+https://github.com/Fak3/mockredis
